# shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 大陆农机配件公共 REST API
  version: 1.0.0
  description: 新增移动端与供求接口的语言无关契约；既有 /api 接口保持兼容。
servers:
  - url: /
paths:
...
  /api/v1/me/notifications:
    get:
      operationId: listNotifications
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/me/notifications/read-all:
    put:
      operationId: readAllNotifications
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/me/notifications/{id}/read:
    put:
      operationId: readNotification
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
    cookieAuth: { type: apiKey, in: cookie, name: dlc_session }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
  responses:
    Success:
      description: 成功
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: integer }
        message: { type: string }
        data: {}
...